run: cargo clippy

+++ examples/05_types/aliasing.rs
@@ -5,7 +5,7 @@ type U64 = u64;
 
 fn main() {
     // `NanoSecond` = `Inch` = `U64` = `u64`.
-    let nanoseconds: NanoSecond = 5 as u64;
+    let nanoseconds: NanoSecond = 5_u64;
     let inches: Inch = 2 as U64;
 
     // Note that type aliases *don't* provide any extra type safety, because
